edgefastapi/requirements.txt
grpcio==1.35.0
protobuf==3.14.0
grpcio-tools==1.35.0
fastapi==0.83.0
uvicorn==0.20.0
fastapi-utils==0.2.1
jinja2==3.1.2
python-multipart==0.0.5
requests==2.28.2
docker==6.0.1
boto3==1.26.96